chore: classify folder structure
