+++ app/layout.tsx
@@ -1,5 +1,6 @@
 import RecoilRootProvider from '@/libs/shared/context/recoil.context'
 import '@/libs/shared/styles/global.scss'
+import { RootLayoutProps } from '@/libs/shared/types/routers.type'
 
 import UiFooter from '@/libs/footer/ui/footer.ui'
 
@@ -9,11 +10,7 @@ import { META_ROOT } from './_meta'
 
 export const metadata = META_ROOT
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
+export default function RootLayout({ children }: RootLayoutProps) {
   return (
     <html lang="ko" className={PureunJeonnam.className}>
       <body>
