--- libs/chip/shared/types/chip.type.ts
@@ -1,11 +1,0 @@
-import { HTMLAttributes } from 'react'
-
-interface ChipContainerProps {
-  chipList: string[]
-}
-
-interface ChipProps extends HTMLAttributes<HTMLSpanElement> {
-  chipName: string
-}
-
-export type { ChipContainerProps, ChipProps }
