+++ libs/dots/ui/dots.ui.tsx
@@ -2,13 +2,11 @@ import classNames from 'classnames/bind'
 
 import Link from 'next/link'
 
+import { UiDotsProps } from '@/libs/shared/types/components.type'
+
 import styles from './dots.module.scss'
 
 const cx = classNames.bind(styles)
-
-interface UiDotsProps {
-  sectionType: 'archives' | 'projects'
-}
 
 export default function UiDots({ sectionType }: UiDotsProps) {
   return (
