+++ libs/gnb/feature/gnb.feature.tsx
@@ -2,14 +2,12 @@
 
 import classNames from 'classnames/bind'
 
+import { GnbProps } from '@/libs/shared/types/components.type'
+
 import useGnbView from '../data-access/gnb.data-access'
 import styles from '../ui/gnb.module.scss'
 
 const cx = classNames.bind(styles)
-
-interface GnbProps {
-  title: string
-}
 
 export default function Gnb({ title }: GnbProps) {
   const { visible, position } = useGnbView()
