--- libs/heading3/shared/heading3.type.ts
@@ -1,8 +1,0 @@
-import { HTMLAttributes } from 'react'
-
-interface HeadingProps extends HTMLAttributes<HTMLHeadingElement> {
-  children?: React.ReactNode
-  color?: string
-}
-
-export type { HeadingProps }
