+++ libs/heroMain/feature/heroMainProjects.feature.tsx
@@ -1,6 +1,7 @@
+import { HeroMainSectionProps } from '@/libs/shared/types/components.type.js'
+
 import UiDots from '@/libs/dots/ui/dots.ui'
 
-import { HeroMainSectionProps } from '../shared/types/heroMain.type'
 import UiHeroMainSection from '../ui/heroMainSection.ui.tsx'
 
 export default function HeroMainProjects({ section }: HeroMainSectionProps) {
