--- libs/heroMain/shared/types/heroMain.type.ts
@@ -1,19 +1,0 @@
-interface HeroMainProps {
-  children: React.ReactNode
-  location: React.ReactElement
-}
-
-interface SectionProps {
-  slug: string
-  title: string
-}
-
-interface SectionTypeProps {
-  sectionType: 'projects' | 'archives'
-}
-
-interface HeroMainSectionProps {
-  section: SectionProps[]
-}
-
-export type { HeroMainProps, SectionTypeProps, HeroMainSectionProps }
