+++ libs/listMain/ui/listMain/index.tsx
@@ -1,5 +1,7 @@
 import classNames from 'classnames/bind'
 
+import { UiListMainProps } from '@/libs/shared/types/components.type'
+
 import UiDividing from '@/libs/dividing/ui/dividing.ui'
 import Heading3 from '@/libs/heading3/ui/heading3.ui'
 import UiSpacing from '@/libs/spacing/ui/spacing.ui'
@@ -7,12 +9,6 @@ import UiSpacing from '@/libs/spacing/ui/spacing.ui'
 import styles from './listMain.module.scss'
 
 const cx = classNames.bind(styles)
-
-interface UiListMainProps {
-  children: React.ReactNode
-  listTitle: 'projects' | 'archives'
-  location: React.ReactElement
-}
 
 export default function UiListMain({
   children,
