--- libs/paragraph/shared/types/paragraph.type.ts
@@ -1,10 +1,0 @@
-import { HTMLAttributes } from 'react'
-
-interface ParagraphProps extends HTMLAttributes<HTMLSpanElement> {
-  children: React.ReactNode
-  color?: string
-  fontWeight?: number
-  fontSize?: number
-}
-
-export type { ParagraphProps }
