+++ libs/shared/_deprecated/components/post/data-access/countHeadings.data-access.ts
