+++ libs/shared/_deprecated/components/post/shared/types/post.type.ts
