+++ libs/shared/_deprecated/types/posts.type.ts
