+++ libs/shared/context/recoil.context.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 
 import { RecoilRoot } from 'recoil'
 
-import { RecoilRootProviderProps } from '../types/recoil.type'
+import { RecoilRootProviderProps } from '@/libs/shared/types/context.type'
 
 export default function RecoilRootProvider({
   children,
